allow the creation of a composition of attributes in one request (#23)
* fix logging to print errors

* minor changes in logging and tests config

* allow to create multiple attribute at once and link them

* make sure tests do not run on outdated app when webpack failed to update srcs

* handle undefined result when fetching data

* make sure the latest version is running as server

* add typing for the composition creation method

* fix test

+++ bin/test.sh
@@ -1,12 +1,21 @@
 #!/usr/bin/env bash
+set -e
 
 export DISABLE_AUTHENTICATION=true
 export NODE_ENV=production
 
 npx webpack --config ./specs.wdio/testapp/webpack.config.js
-node --enable-source-maps specs.wdio/testapp/testserver.js&
+bin/testserver.sh&
 SERVER_PID=$!
+
+sleep 2
+
+if ! ps -p $SERVER_PID; then
+  echo "Node test server is NOT running!!! Maybe the server is running already and blocking the port. Check for 'ps aux | grep testserver' and terminate it."
+  exit 1
+fi
+
 echo "Node server is running on with PID ${SERVER_PID}"
-sleep 1
+
 npx karma start
 kill $SERVER_PID
