+++ bin/test.wdio.sh
@@ -1,4 +1,5 @@
 #!/usr/bin/env bash
+set -e
 
 export NODE_ENV=production
 export SERVER_BASE_URL=http://localhost:3000
@@ -8,8 +9,16 @@ npx webpack --config ./specs.wdio/testapp/webpack.config.js
 
 bin/testserver.sh&
 SERVER_PID=$!
-echo "Node server is running on with PID ${SERVER_PID}"
+
 sleep 2
+
+if ! ps -p $SERVER_PID; then
+  echo "Node test server is NOT running!!! Maybe the server is running already and blocking the port. Check for 'ps aux | grep testserver' and terminate it."
+  exit 1
+fi
+
+echo "Node server is running on with PID ${SERVER_PID}"
+
 npx wdio run ./wdio.conf.ts
 succ=$?
 kill $SERVER_PID
